% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customConverters.R
\name{as.DISCBIO}
\alias{as.DISCBIO}
\title{Convert Single Cell Data Objects to DISCBIO.}
\usage{
as.DISCBIO(x, ...)
}
\arguments{
\item{x}{an object of class Seurat or SingleCellExperiment}

\item{...}{additional parameters to pass to the function}
}
\value{
a DISCBIO-class object
}
\description{
Initialize a DISCBIO-class object starting from a
SingleCellExperiment object or a Seurat object
}
\details{
Additional parameters to pass to `list` include, if x is
  a Seurat object, "assay", which is a string indicating the assay slot
  used to obtain data from (defaults to 'RNA')
}
