% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samr-adapted.R
\name{check.format}
\alias{check.format}
\title{Check format}
\usage{
check.format(y, resp.type, censoring.status = NULL)
}
\arguments{
\item{y}{y}

\item{resp.type}{resp type}

\item{censoring.status}{censoring status}
}
\description{
Check format
}
