% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassVectoringDT.R
\name{ClassVectoringDT}
\alias{ClassVectoringDT}
\title{title}
\usage{
ClassVectoringDT(object, Cluster_ID, K, First, Second, sigDEG)
}
\arguments{
\item{object}{object}

\item{Cluster_ID}{Cluster_ID}

\item{K}{K}

\item{First}{First}

\item{Second}{Second}

\item{sigDEG}{sigDEG}
}
\description{
description
}
