% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscBIO_class.R
\name{clustfun}
\alias{clustfun}
\title{title}
\usage{
clustfun(
  x,
  clustnr = 20,
  bootnr = 50,
  metric = "pearson",
  do.gap = TRUE,
  SE.method = "Tibs2001SEmax",
  SE.factor = 0.25,
  B.gap = 50,
  cln = 0,
  rseed = 17000
)
}
\arguments{
\item{x}{x}

\item{clustnr}{clustnr}

\item{bootnr}{bootnr}

\item{metric}{metric}

\item{do.gap}{do.gap}

\item{SE.method}{SE.method}

\item{SE.factor}{SE.factor}

\item{B.gap}{B.gap}

\item{cln}{cln}

\item{rseed}{rseed}
}
\description{
description
}
