% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscBIO_class.R
\name{comptSNE,PSCANseq -method}
\alias{comptSNE,PSCANseq -method}
\title{title}
\usage{
\S4method{comptSNE}{`PSCANseq `}(object, rseed = 15555)
}
\arguments{
\item{object}{object}

\item{rseed}{rseed}
}
\description{
description
}
