% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-generic-comptSNE.R
\name{comptSNE}
\alias{comptSNE}
\title{Computing tSNE for K-means clustering}
\usage{
comptSNE(object, rseed = 15555, quiet = FALSE)
}
\arguments{
\item{object}{\code{PSCANseq} class object.}

\item{quiet}{if `TRUE`, suppresses intermediate output}

\item{Integer}{number. Random seed to to yield exactly reproducible maps across different runs. Default is 15555.}
}
\description{
This function is used to compute the t-Distributed Stochastic Neighbor Embedding (t-SNE).
}
