% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.val.R
\name{cross.val}
\alias{cross.val}
\title{title}
\usage{
cross.val(exp.df, class.vec, segments, performance, class.algo)
}
\arguments{
\item{exp.df}{exp.df}

\item{class.vec}{class.vec}

\item{segments}{segments}

\item{performance}{performance}

\item{class.algo}{class.algo}
}
\description{
description
}
