% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-generic-DEGanalysis.R, R/VolcanoPlot.R
\name{DEGanalysis}
\alias{DEGanalysis}
\alias{VolcanoPlot}
\title{Determining differentially expressed genes (DEGs) between all individual clusters.}
\usage{
DEGanalysis(
  object,
  Clustering = "K-means",
  K,
  fdr = 0.05,
  name = "Name",
  export = TRUE
)

VolcanoPlot(object, value = 0.05, name, fc = 0.5, adj = TRUE, FS = 0.4)
}
\arguments{
\item{object}{A data frame showing the differentially expressed genes (DEGs) in a particular cluster}

\item{Clustering}{Clustering has to be one of the following: ["K-means","MB"]. Default is "K-means"}

\item{K}{A numeric value of the number of clusters.}

\item{fdr}{A numeric value of the false discovery rate. Default is 0.05.}

\item{name}{A string vector showing the name to be used to save the resulted tables.}

\item{export}{A logical vector that allows writing the final gene list in excel file. Default is TRUE.}

\item{fc}{A numeric value of the fold change. Default is 0.5.}

\item{adj}{A logical vector that allows adjusting the y axis by adding a very smal value (0.00000000001) to logFDR. Default is TRUE.}

\item{FS}{A numeric value of the font size. Default is 0.4.}

\item{Value}{A numeric value of the false discovery rate. Default is 0.05.. Default is 0.05}
}
\description{
This function defines DEGs between all individual clusters generated by either K-means or model based clustering.

Plotting differentially expressed genes (DEGs) in a particular cluster. 
Volcano plots are used to readily show the DEGs by plotting significance versus fold-change on the y and x axes, respectively.
}
