% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-generic-DEGanalysisM.R
\name{DEGanalysisM}
\alias{DEGanalysisM}
\title{Determining differentially expressed genes (DEGs) between two particular clusters.}
\usage{
DEGanalysisM(
  object,
  Clustering = "K-means",
  K,
  fdr = 0.05,
  name = "Name",
  First = "CL1",
  Second = "CL2",
  export = TRUE
)
}
\arguments{
\item{object}{\code{PSCANseq} class object.}

\item{Clustering}{Clustering has to be one of the following: ["K-means","MB"]. Default is "K-means"}

\item{K}{A numeric value of the number of clusters.}

\item{fdr}{A numeric value of the false discovery rate. Default is 0.05.}

\item{name}{A string vector showing the name to be used to save the resulted tables.}

\item{First}{A string vector showing the first target cluster.  Default is "CL1"}

\item{Second}{A string vector showing the second target cluster.  Default is "CL2"}

\item{export}{A logical vector that allows writing the final gene list in excel file. Default is TRUE.}
}
\description{
This function defines DEGs between particular clusters generated by either K-means or model based clustering.
}
