% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customConverters.R
\name{DISCBIO2SingleCellExperiment}
\alias{DISCBIO2SingleCellExperiment}
\title{Convert a DISCBIO object to a SingleCellExperiment.}
\usage{
DISCBIO2SingleCellExperiment(x)
}
\arguments{
\item{x}{an object of class DISCBIO}
}
\value{
a SingleCellExperiment-class object
}
\description{
Extract the SingleCellExperiment input data from the corresponding input slot
in a DISCBIO-class object
}
\examples{
g1_disc <- DISCBIO(valuesG1msTest)
class(g1_disc)
g1_sce <- DISCBIO2SingleCellExperiment(g1_disc)
class(g1_sce)


}
