% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscBIO_class.R
\name{estimateSizeFactorsForMatrix}
\alias{estimateSizeFactorsForMatrix}
\title{title}
\usage{
estimateSizeFactorsForMatrix(counts, locfunc = median)
}
\arguments{
\item{counts}{counts}

\item{locfunc}{locfunc}
}
\description{
description
}
