% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscBIO_class.R
\name{Exprmclust}
\alias{Exprmclust}
\title{title}
\usage{
Exprmclust(
  data,
  clusternum = 2:9,
  modelNames = "VVV",
  reduce = T,
  cluster = NULL
)
}
\arguments{
\item{data}{data}

\item{clusternum}{clusternum}

\item{modelNames}{modelNames}

\item{reduce}{reduce}

\item{cluster}{cluster}
}
\description{
description
}
