% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-generic-FinalPreprocessing.R
\name{FinalPreprocessing}
\alias{FinalPreprocessing}
\alias{FinalPreprocessing,PSCANseq-method}
\title{Final Preprocessing}
\usage{
FinalPreprocessing(object, GeneFlitering = "NoiseF", export = TRUE)

\S4method{FinalPreprocessing}{PSCANseq}(object, GeneFlitering = "NoiseF", export = TRUE)
}
\arguments{
\item{object}{\code{PSCANseq} class object.}

\item{GeneFlitering}{GeneFlitering has to be one of the followings: ["NoiseF","ExpF"]. Default is "NoiseF"}

\item{export}{A logical vector that allows writing the final gene list in excel file. Default is TRUE.}
}
\description{
This function generates the final filtered normalized dataset.
}
