% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{HumanMouseGeneIds}
\alias{HumanMouseGeneIds}
\title{Human and Mouse Gene Identifiers.}
\source{
Data were imported, modified, and formatted from the
Mus.musculus (ver 1.3.1) and the Homo.sapiens (ver 1.3.1)
BioConductor libraries.
}
\description{
Data.frame including ENTREZID, SYMBOL, and ENSEMBL gene identifiers
of human and mouse genes.
}
