% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO.R, R/DiscBIO_class.R
\name{initialize,PSCANseq -method}
\alias{initialize,PSCANseq -method}
\title{title}
\usage{
\S4method{initialize}{`PSCANseq `}(.Object, expdata)

\S4method{initialize}{`PSCANseq `}(.Object, expdata)
}
\arguments{
\item{.Object}{.Object}

\item{expdata}{expdata}
}
\description{
description

description
}
