% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J48DT.R
\name{J48DT}
\alias{J48DT}
\title{J48 Decision Tree}
\usage{
J48DT(data, quiet = FALSE, plot = TRUE)
}
\arguments{
\item{data}{A data frame resulted from running the function ClassVectoringDT.}

\item{quiet}{If `TRUE`, suppresses intermediary output}

\item{plot}{If `FALSE`, suppresses plot output}
}
\value{
Information about the J48 model and, by default, a plot of the
  decision tree.
}
\description{
The decision tree analysis is implemented over a training
  dataset, which consisted of the DEGs obtained by either SAMseq or the
  binomial differential expression.
}
\examples{
data(valuesG1msReduced_treated_K)  # details: DIscBIO:::prepExampleDataset
sc <- valuesG1msReduced_treated_K
cdiff <- DEGanalysis2clust(
    sc, Clustering="K-means", K=3, fdr=.2, name="Name", First="CL1",
    Second="CL2", export=FALSE,
    nresamp=5, nperms=20
)
sigDEG <- cdiff[[1]]
DATAforDT <- ClassVectoringDT(
    sc, Clustering="K-means", K=3, First="CL1", Second="CL2", sigDEG,
)
J48DT(DATAforDT)
}
