% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J48DTeval.R
\name{J48DTeval}
\alias{J48DTeval}
\title{Evaluating the performance of the J48 decision tree.}
\usage{
J48DTeval(data, num.folds = 10, First = "CL1", Second = "CL2", quiet = FALSE)
}
\arguments{
\item{data}{The resulted data from running the function J48DT.}

\item{num.folds}{A numeric value of the number of folds for the cross
validation assessment. Default is 10.}

\item{First}{A string vector showing the first target cluster.  Default is
"CL1"}

\item{Second}{A string vector showing the second target cluster.  Default is
"CL2"}

\item{quiet}{If `TRUE`, suppresses intermediary output}
}
\value{
Statistics about the J48 model
}
\description{
This function evaluates the performance of the generated trees
  for error estimation by ten-fold cross validation assessment.
}
