% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jaccard.R
\name{Jaccard}
\alias{Jaccard}
\title{title}
\usage{
Jaccard(object, Clustering, K, plot = TRUE)
}
\arguments{
\item{object}{object}

\item{Clustering}{Clustering}

\item{K}{K}

\item{plot}{if `TRUE`, plots the mean Jaccard similarities}
}
\description{
description
}
