% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-generic-KMClustDiffGenes.R
\name{KMClustDiffGenes}
\alias{KMClustDiffGenes}
\alias{KMClustDiffGenes,DISCBIO-method}
\title{ClustDiffGenes}
\usage{
KMClustDiffGenes(object, K, fdr = 0.01)

\S4method{KMClustDiffGenes}{DISCBIO}(object, K, fdr = 0.01)
}
\arguments{
\item{object}{\code{DISCBIO} class object.}

\item{K}{A numeric value of the number of clusters.}

\item{fdr}{A numeric value of the false discovery rate. Default is 0.01.}
}
\description{
description
}
\examples{
sc <- DISCBIO(valuesG1ms)
sc <- Clustexp(sc, cln=3, quiet=TRUE) # K-means clustering
sc <- comptSNE(sc, rseed=15555, quiet=TRUE)
 KMClustDiffGenes(sc, K=3, fdr=.3)
}
