% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscBIO_class.R
\name{KMclustheatmap,PSCANseq -method}
\alias{KMclustheatmap,PSCANseq -method}
\title{title}
\usage{
\S4method{KMclustheatmap}{`PSCANseq `}(object, hmethod = "single")
}
\arguments{
\item{object}{object}

\item{hmethod}{hmethod}
}
\description{
description
}
