% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-generic-KMclustheatmap.R
\name{KMclustheatmap}
\alias{KMclustheatmap}
\alias{KMclustheatmap,PSCANseq-method}
\title{title}
\usage{
KMclustheatmap(object, hmethod = "single", plot = TRUE)

\S4method{KMclustheatmap}{PSCANseq}(object, hmethod = "single", plot = TRUE)
}
\arguments{
\item{object}{object}

\item{hmethod}{hmethod}

\item{plot}{if `TRUE`, plots the heatmap; otherwise, just prints cclmo}
}
\description{
description
}
