% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KmeanOrder.R
\name{KmeanOrder}
\alias{KmeanOrder}
\title{title}
\usage{
KmeanOrder(object, Clusters, sampleNames, quiet = FALSE, export = TRUE)
}
\arguments{
\item{object}{object}

\item{Clusters}{Clusters}

\item{sampleNames}{sampleNames}

\item{quiet}{if `TRUE`, suppresses intermediary output}

\item{export}{if `TRUE`, exports order table to csv}
}
\description{
description
}
