% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KmeansCBI.R
\name{KmeansCBI}
\alias{KmeansCBI}
\title{K-means CBI}
\usage{
KmeansCBI(
  data,
  krange,
  k = NULL,
  scaling = FALSE,
  runs = 1,
  criterion = "ch",
  method = "euclidean",
  ...
)
}
\arguments{
\item{data}{data}

\item{krange}{krange}

\item{k}{k}

\item{scaling}{scaling}

\item{runs}{runs}

\item{criterion}{criterion}

\item{method}{method}

\item{...}{extra parameters to be passed to kmm through Kmeansruns}
}
\description{
K-means CBI
}
