% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kmeansruns.R
\name{Kmeansruns}
\alias{Kmeansruns}
\title{K-means runs}
\usage{
Kmeansruns(
  data,
  krange = 2:10,
  criterion = "ch",
  iter.max = 100,
  runs = 100,
  scaledata = FALSE,
  alpha = 0.001,
  critout = FALSE,
  plot = FALSE,
  method = "euclidean",
  ...
)
}
\arguments{
\item{runs}{number of runs}

\item{...}{extra parameters to be passed to kmm}
}
\description{
K-means runs
}
