% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscBIO_class.R
\name{MB_Order}
\alias{MB_Order}
\title{title}
\usage{
MB_Order(object, sampleNames, Names)
}
\arguments{
\item{object}{object}

\item{sampleNames}{sampleNames}

\item{Names}{Names}
}
\description{
description
}
