% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-generic-MBClustDiffGenes.R,
%   R/MBClustDiffGenes.R
\name{MBClustDiffGenes}
\alias{MBClustDiffGenes}
\alias{MBClustDiffGenes,DISCBIO-method}
\title{ClustDiffGenes}
\usage{
MBClustDiffGenes(object, K, fdr = 0.01, export = TRUE, quiet = FALSE)

\S4method{MBClustDiffGenes}{DISCBIO}(object, K, fdr = 0.01, export = TRUE, quiet = FALSE)

MBClustDiffGenes(object, K, fdr = 0.01, export = TRUE, quiet = FALSE)

\S4method{MBClustDiffGenes}{DISCBIO}(object, K, fdr = 0.01, export = TRUE, quiet = FALSE)
}
\arguments{
\item{object}{\code{DISCBIO} class object.}

\item{K}{A numeric value of the number of clusters.}

\item{fdr}{A numeric value of the false discovery rate. Default is 0.01.}

\item{export}{A logical vector that allows writing the final gene list in excel file. Default is TRUE.}

\item{quiet}{if `TRUE`, suppresses intermediate text output}
}
\description{
ClustDiffGenes

ClustDiffGenes
}
