% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-generic-MBclustheatmap.R
\name{MBclustheatmap}
\alias{MBclustheatmap}
\alias{MBclustheatmap,PSCANseq-method}
\title{title}
\usage{
MBclustheatmap(object, hmethod = "single")

\S4method{MBclustheatmap}{PSCANseq}(object, hmethod = "single")
}
\arguments{
\item{object}{object}

\item{hmethod}{hmethod}
}
\description{
title
}
