% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetAnalysis.R
\name{NetAnalysis}
\alias{NetAnalysis}
\title{Networking analysis.}
\usage{
NetAnalysis(data)
}
\arguments{
\item{data}{Protein-protein interaction data frame resulted from running the PPI function.}
}
\description{
This function checks the connectivity degree and the betweenness centrality, which reflect the communication flow in the defined PPI networks
}
