% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networking.R
\name{Networking}
\alias{Networking}
\title{Plotting the network.}
\usage{
Networking(data, FileName, species = "9606")
}
\arguments{
\item{data}{A gene list.}

\item{FileName}{A string vector showing the name to be used to save the resulted network.}

\item{species}{The taxonomy name/id. Default is "9606" for Homo sapiens.}
}
\description{
This function uses STRING-api to plot the network.
}
