% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscBIO_class.R
\name{NoiseFiltering}
\alias{NoiseFiltering}
\title{title}
\usage{
NoiseFiltering(
  object,
  percentile,
  CV,
  GeneList,
  geneCol,
  FgeneCol,
  erccCol,
  Val = TRUE
)
}
\arguments{
\item{object}{object}

\item{percentile}{percentile}

\item{CV}{CV}

\item{GeneList}{GeneList}

\item{geneCol}{geneCol}

\item{FgeneCol}{FgeneCol}

\item{erccCol}{erccCol}

\item{Val}{Val}
}
\description{
description
}
