% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscBIO_class.R
\name{Normalizedata}
\alias{Normalizedata}
\alias{Normalizedata,PSCANseq-method}
\title{title}
\usage{
Normalizedata(
  object,
  mintotal = 1000,
  minexpr = 0,
  minnumber = 0,
  maxexpr = Inf,
  downsample = FALSE,
  dsn = 1,
  rseed = 17000
)

\S4method{Normalizedata}{PSCANseq}(
  object,
  mintotal = 1000,
  minexpr = 0,
  minnumber = 0,
  maxexpr = Inf,
  downsample = FALSE,
  dsn = 1,
  rseed = 17000
)
}
\arguments{
\item{object}{object}

\item{mintotal}{mintotal}

\item{minexpr}{minexpr}

\item{minnumber}{minnumber}

\item{maxexpr}{maxexpr}

\item{downsample}{downsample}

\item{dsn}{dsn}

\item{rseed}{rseed}
}
\description{
title
}
