% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscBIO_class.R
\name{plotExptSNE}
\alias{plotExptSNE}
\alias{plotExptSNE,PSCANseq-method}
\title{title}
\usage{
plotExptSNE(object, g, n = "")

\S4method{plotExptSNE}{PSCANseq}(object, g, n = "")
}
\arguments{
\item{object}{object}

\item{g}{g}

\item{n}{n}
}
\description{
title
}
