% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-classes.R,
%   R/DIscBIO-generic-plotexptsneMB.R
\name{plotexptsneMB}
\alias{plotexptsneMB}
\alias{plotexptsneMB,PSCANseq-method}
\title{title}
\usage{
plotexptsneMB(object, g, n = "")

\S4method{plotexptsneMB}{PSCANseq}(object, g, n = "")

plotexptsneMB(object, g, n = "")

\S4method{plotexptsneMB}{PSCANseq}(object, g, n = "")
}
\arguments{
\item{object}{object}

\item{g}{g}

\item{n}{n}
}
\description{
title

title
}
