% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscBIO_class.R
\name{plotKmeansLabelstSNE,PSCANseq -method}
\alias{plotKmeansLabelstSNE,PSCANseq -method}
\title{title}
\usage{
\S4method{plotKmeansLabelstSNE}{`PSCANseq `}(object)
}
\arguments{
\item{object}{object}
}
\description{
description
}
