% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-generic-plotLabelstSNE.R
\name{plotLabelstSNE}
\alias{plotLabelstSNE}
\alias{plotLabelstSNE,DISCBIO-method}
\title{tSNE map with labels}
\usage{
plotLabelstSNE(object)

\S4method{plotLabelstSNE}{DISCBIO}(object)
}
\arguments{
\item{object}{\code{DISCBIO} class object.}
}
\value{
Plot containing the ID of the cells in each cluster
}
\description{
Visualizing k-means or model-based clusters using tSNE maps
}
