% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-classes.R,
%   R/DIscBIO-generic-plotMBLabelstSNE.R
\name{plotMBLabelstSNE}
\alias{plotMBLabelstSNE}
\alias{plotMBLabelstSNE,PSCANseq-method}
\title{title}
\usage{
plotMBLabelstSNE(object)

\S4method{plotMBLabelstSNE}{PSCANseq}(object)

plotMBLabelstSNE(object)

\S4method{plotMBLabelstSNE}{PSCANseq}(object)
}
\arguments{
\item{object}{object}
}
\description{
description

description
}
