% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMBorderPCA.R
\name{PlotMBorderPCA}
\alias{PlotMBorderPCA}
\title{Plotting pseudo-time ordering in Model-based clustering in PCA.}
\usage{
PlotMBorderPCA(object)
}
\arguments{
\item{object}{\code{DISCBIO} class object.}
}
\value{
A plot of the PCA.
}
\description{
The PCA representation can also be used to show the pseudo-time
  ordering.
}
\examples{
data(valuesG1msReduced_treated_MB)  # details: DIscBIO:::prepExampleDataset
sc <- valuesG1msReduced_treated_MB
sc <- MB_Order(sc, export = FALSE)
PlotMBorderPCA(sc)
}
