% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMBpca.R
\name{PlotMBpca}
\alias{PlotMBpca}
\title{Plotting pseudo-time ordering or gene expression in Model-based
clustering in PCA}
\usage{
PlotMBpca(object, type = "order", g = NULL, n = NULL)
}
\arguments{
\item{object}{\code{DISCBIO} class object.}

\item{type}{either `order` to plot pseudo-time ordering or `exp` to plot gene
expression}

\item{g}{Individual gene name or vector with a group of gene names
corresponding to a subset of valid row names of the \code{ndata} slot of
the \code{DISCBIO} object. Ignored if `type="order"`.}

\item{n}{String of characters representing the title of the plot. Default is
NULL and the first element of \code{g} is chosen. Ignored if
`type="order"`.}
}
\value{
A plot of the PCA.
}
\description{
The PCA representation can either be used to show pseudo-time
ordering or the gene expression of a particular gene.
}
