% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotmclust.R
\name{Plotmclust}
\alias{Plotmclust}
\title{title}
\usage{
Plotmclust(
  mclustobj,
  x = 1,
  y = 2,
  MSTorder = NULL,
  show_tree = T,
  show_full_tree = F,
  show_cell_names = F,
  cell_name_size = 3,
  markerexpr = NULL,
  showcluster = T
)
}
\arguments{
\item{mclustobj}{mclustobj}

\item{x}{x}

\item{y}{y}

\item{MSTorder}{MSTorder}

\item{show_tree}{show_tree}

\item{show_full_tree}{show_full_tree}

\item{show_cell_names}{show_cell_names}

\item{cell_name_size}{cell_name_size}

\item{markerexpr}{markerexpr}

\item{showcluster}{showcluster}
}
\description{
description
}
