% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotmclustMB.R
\name{PlotmclustMB}
\alias{PlotmclustMB}
\title{Plotting the Model-based clusters in PCA.}
\usage{
PlotmclustMB(object)
}
\arguments{
\item{object}{\code{PSCANseq} class object.}
}
\description{
Plot the model-based clustering results
}
