% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOrderKMtsne.R
\name{plotOrderKMtsne}
\alias{plotOrderKMtsne}
\title{Plotting the pseudo-time ordering based on K-means clusters in the t-SNE map}
\usage{
plotOrderKMtsne(object)
}
\arguments{
\item{object}{\code{PSCANseq} class object.
of the \code{PSCANseq} object.}
}
\description{
The tSNE representation can also be used to show the pseudo-time ordering.
}
