% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-generic-plotOrderMBtsne.R
\name{plotOrderMBtsne}
\alias{plotOrderMBtsne}
\alias{plotOrderMBtsne,DISCBIO-method}
\title{Plotting the pseudo-time ordering based on Model-based clusters in the
  t-SNE map}
\usage{
plotOrderMBtsne(object)

\S4method{plotOrderMBtsne}{DISCBIO}(object)
}
\arguments{
\item{object}{\code{DISCBIO} class object.}
}
\value{
A plot of the pseudo-time ordering.
}
\description{
The tSNE representation can also be used to show the pseudo-time
  ordering.
}
\examples{
sc<- DISCBIO(valuesG1msReduced)
sc<-NoiseFiltering(sc,percentile=0.9, CV=0.2, export=FALSE)
sc<-Normalizedata(sc)
sc<-FinalPreprocessing(sc,GeneFlitering="NoiseF", export=FALSE)
sc <- Exprmclust(sc, K=2, reduce=TRUE, quiet=TRUE)
sc<- comptsneMB(sc, rseed=15555, quiet=TRUE, max_iter=100)
plotOrderMBtsne(sc)
}
