% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-generic-plotOrderTsne.R
\name{plotOrderTsne}
\alias{plotOrderTsne}
\alias{plotOrderTsne,DISCBIO-method}
\title{Plotting the pseudo-time ordering in the t-SNE map}
\usage{
plotOrderTsne(object)

\S4method{plotOrderTsne}{DISCBIO}(object)
}
\arguments{
\item{object}{\code{DISCBIO} class object.}
}
\value{
A plot of the pseudo-time ordering.
}
\description{
The tSNE representation can also be used to show the pseudo-time
  ordering.
}
