% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-generic-plottSNE.R
\name{plottSNE}
\alias{plottSNE}
\title{tSNE map for K-means clustering}
\usage{
plottSNE(object)
}
\arguments{
\item{object}{\code{DISCBIO} class object.}
}
\description{
Visualizing the K-means clusters using tSNE maps
}
