% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-generic-plottsneMB.R
\name{plottsneMB}
\alias{plottsneMB}
\alias{plottsneMB,PSCANseq-method}
\title{tSNE map for Model-based clustering}
\usage{
plottsneMB(object, K)

\S4method{plottsneMB}{PSCANseq}(object, K)
}
\arguments{
\item{object}{\code{PSCANseq} class object.}

\item{K}{A numeric value of the number of clusters}
}
\description{
Visualizing the Model-based clusters using tSNE maps
}
