% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPI.R
\name{PPI}
\alias{PPI}
\title{Defining protein-protein interactions (PPI) over a list of genes,}
\usage{
PPI(data, FileName = NULL, species = "9606")
}
\arguments{
\item{data}{A gene list.}

\item{FileName}{A string vector showing the name to be used to save the
resulted table. If null, no file will be exported}

\item{species}{The taxonomy name/id. Default is "9606" for Homo sapiens.}
}
\value{
Either a TSV file stored in the user's file system and its
corresponding `data.frame` object in R or and R object containing that
information.
}
\description{
This function uses STRING-api. The outcome of STRING analysis
  will be stored in tab separated values (TSV) files.
}
\examples{
\donttest{
sc <- DISCBIO(valuesG1msReduced)
sc <- NoiseFiltering(sc, percentile=0.9, CV=0.2, export=FALSE)
sc <- Normalizedata(
    sc, mintotal=1000, minexpr=0, minnumber=0, maxexpr=Inf, downsample=FALSE,
    dsn=1, rseed=17000
)
sc <- FinalPreprocessing(sc, export=FALSE, quiet=TRUE)
sc <- Clustexp(sc, cln=3, quiet=TRUE) # K-means clustering
sc <- comptSNE(sc, max_iter=100, quiet=TRUE)
dff <- DEGanalysis2clust(sc, K=3, export=FALSE, quiet=TRUE, plot=FALSE)
PPI(dff$FinalDEGsU[, "Gene Name"])
}
}
