% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepExampleDataset.R
\name{prepExampleDataset}
\alias{prepExampleDataset}
\title{Prepare Example Dataset}
\usage{
prepExampleDataset(clustering)
}
\arguments{
\item{clustering}{determines the clustering method (either "K" or "MB").}
}
\value{
rda file
}
\description{
Internal function that prepares a pre-treated dataset for use in
several examples
}
