% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-classes.R
\docType{class}
\name{PSCANseq}
\alias{PSCANseq}
\alias{PSCANseq-class,}
\alias{PSCANseq-class}
\title{The PSCANseq Class}
\arguments{
\item{object}{An PSCANseq object.}
}
\description{
The PSCANseq class is the central object storing all information generated throughout the pipeline.
}
\details{
PSCANseq
}
\section{Slots}{

\describe{
\item{\code{expdata}}{<slot description>}

\item{\code{expdataAll}}{<slot description>}

\item{\code{ndata}}{<slot description>}

\item{\code{fdata}}{<slot description>}

\item{\code{distances}}{<slot description>}

\item{\code{tsne}}{<slot description>}

\item{\code{background}}{<slot description>}

\item{\code{out}}{<slot description>}

\item{\code{cpart}}{<slot description>}

\item{\code{fcol}}{<slot description>}

\item{\code{filterpar}}{<slot description>}

\item{\code{clusterpar}}{<slot description>}

\item{\code{outlierpar}}{<slot description>}

\item{\code{kmeans}}{<slot description>}

\item{\code{MBclusters}}{<slot description>}

\item{\code{kordering}}{<slot description>}

\item{\code{MBordering}}{<slot description>}

\item{\code{MBtsne}}{<slot description>}

\item{\code{noiseF}}{<slot description>}

\item{\code{FinalGeneList}}{<slot description>}
}}

