% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-generic-pseudoTimeOrdering.R
\name{pseudoTimeOrdering}
\alias{pseudoTimeOrdering}
\alias{pseudoTimeOrdering,DISCBIO-method}
\title{Pseudo-time ordering}
\usage{
pseudoTimeOrdering(
  object,
  quiet = FALSE,
  export = FALSE,
  filename = "Cellular_pseudo-time_ordering"
)

\S4method{pseudoTimeOrdering}{DISCBIO}(
  object,
  quiet = FALSE,
  export = FALSE,
  filename = "Cellular_pseudo-time_ordering"
)
}
\arguments{
\item{object}{\code{DISCBIO} class object.}

\item{quiet}{if `TRUE`, suppresses intermediary output}

\item{export}{if `TRUE`, exports order table to csv}

\item{filename}{Name of the exported file (if `export=TRUE`)}
}
\value{
The DISCBIO-class object input with the kordering slot filled.
}
\description{
This function takes the exact output of exprmclust function and
  construct Pseudo-time ordering by mapping all cells onto the path that
  connects cluster centers.
}
