% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{reformatSiggenes}
\alias{reformatSiggenes}
\title{Reformat Siggenes Table}
\usage{
reformatSiggenes(table)
}
\arguments{
\item{table}{output from `samr::samr.compute.siggenes.table`}
}
\description{
Reformats the Siggenes table output from the SAMR package
}
\seealso{
replaceDecimals
}
\author{
Waldir Leoncio
}
