% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceDecimals.R
\name{replaceDecimals}
\alias{replaceDecimals}
\title{Replace Decimals}
\usage{
replaceDecimals(x, from = ",", to = ".")
}
\description{
Replaces decimals separators between comma and periods
}
\note{
This function was especially designed to be used with retormatSiggenes
}
\seealso{
reformatSiggenes
}
