% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions.R
\name{retrieveURL}
\alias{retrieveURL}
\title{Retries a URL}
\usage{
retrieveURL(data, species, outputFormat, maxRetries = 3, successCode = 200)
}
\arguments{
\item{maxRetries}{maximum number of attempts to connect to the STRING api.}

\item{successCode}{Status code number that represents success}
}
\value{
either the output of httr::GET or an error message
}
\description{
Retries a URL
}
\author{
Waldir Leoncio
}
