% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RpartEVAL.R
\name{RpartEVAL}
\alias{RpartEVAL}
\title{Evaluating the performance of the RPART Decision Tree.}
\usage{
RpartEVAL(data, num.folds = 10, First = "CL1", Second = "CL2", quiet = FALSE)
}
\arguments{
\item{data}{The resulted data from running the function J48DT.}

\item{num.folds}{A numeric value of the number of folds for the cross validation assessment. Default is 10.}

\item{First}{A string vector showing the first target cluster.  Default is "CL1"}

\item{Second}{A string vector showing the second target cluster.  Default is "CL2"}

\item{quiet}{If `TRUE`, suppresses intermediary output}
}
\description{
This function evaluates the performance of the generated trees for error estimation by ten-fold cross validation assessment.
}
\examples{
\dontrun{
sc <- DISCBIO(valuesG1msReduced)
sc <- NoiseFiltering(sc, export=FALSE)
sc <- Normalizedata(
    sc, mintotal=1000, minexpr=0, minnumber=0, maxexpr=Inf, downsample=FALSE,
    dsn=1, rseed=17000
)
sc <- FinalPreprocessing(sc, GeneFlitering="NoiseF", export=FALSE)
sc <- Clustexp(sc, cln=3) # K-means clustering
sc <- comptSNE(sc, rseed=15555)
cdiff <- DEGanalysis2clust(
    sc, Clustering="K-means", K=3, fdr=.2, name="Name", First="CL1",
    Second="CL2", export=FALSE
)
sigDEG <- cdiff[[1]]
DATAforDT <- ClassVectoringDT(
    sc, Clustering="K-means", K=3, First="CL1", Second="CL2", sigDEG,
)
RpartEVAL(DATAforDT,num.folds=10,First="CL1",Second="CL2")
}
}
