% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samr-adapted.R
\name{samr.estimate.depth}
\alias{samr.estimate.depth}
\title{Estimate sequencing depths}
\usage{
samr.estimate.depth(x)
}
\arguments{
\item{x}{data matrix. nrow=#gene, ncol=#sample}
}
\value{
depth: estimated sequencing depth. a vector with len sample.
}
\description{
Estimate sequencing depths
}
