% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{valuesG1ms}
\alias{valuesG1ms}
\title{valuesG1ms A dataset of 94 single cells from a myxoid liposarcoma cell line. 
Columns refer to samples and rows refer to genes. The last 92 rows refer to external RNA controls consortium (ERCC) spike-ins}
\description{
valuesG1ms A dataset of 94 single cells from a myxoid liposarcoma cell line. 
Columns refer to samples and rows refer to genes. The last 92 rows refer to external RNA controls consortium (ERCC) spike-ins
}
