% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{valuesG1msReduced}
\alias{valuesG1msReduced}
\title{valuesG1msReduced A dataset of 30 single cells from a myxoid liposarcoma cell line.
Columns refer to samples and rows refer to genes. The last 92 rows refer to external RNA controls consortium (ERCC) spike-ins.}
\description{
valuesG1msReduced A dataset of 30 single cells from a myxoid liposarcoma cell line.
Columns refer to samples and rows refer to genes. The last 92 rows refer to external RNA controls consortium (ERCC) spike-ins.
}
