% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{valuesG1msReduced_treated_K}
\alias{valuesG1msReduced_treated_K}
\title{Single-cells data from a myxoid liposarcoma cell line (K variant)}
\description{
Equivalent to valuesG1msReduced after treatment and K-means clustering.
}
