% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{valuesG1msReduced_treated_MB}
\alias{valuesG1msReduced_treated_MB}
\title{Single-cells data from a myxoid liposarcoma cell line (MB variant)}
\description{
Equivalent to valuesG1msReduced after treatment and model-based clustering.
}
