% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{valuesG1msTest}
\alias{valuesG1msTest}
\title{Single-cells data from a myxoid liposarcoma cell line}
\description{
A sample of single cells from a myxoid liposarcoma cell
  line. Columns refer to samples and rows refer to genes. The last rows
  refer to external RNA controls consortium (ERCC) spike-ins. This dataset is
  part of a larger dataset containing 94 single cells. The complete dataset
  is fully compatible with this package and an rda file can be obtained at
  https://github.com/ocbe-uio/DIscBIO/blob/dev/data/valuesG1ms.rda
}
