% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VolcanoPlot.R
\name{VolcanoPlot}
\alias{VolcanoPlot}
\title{Volcano Plot}
\usage{
VolcanoPlot(object, value = 0.05, name, fc = 0.5, FS = 0.4)
}
\arguments{
\item{object}{A data frame showing the differentially expressed genes (DEGs)
in a particular cluster}

\item{value}{A numeric value of the false discovery rate. Default is 0.05..
Default is 0.05}

\item{name}{A string vector showing the name to be used to save the resulted
tables.}

\item{fc}{A numeric value of the fold change. Default is 0.5.}

\item{FS}{A numeric value of the font size. Default is 0.4.}
}
\value{
A volcano plot
}
\description{
Plotting differentially expressed genes (DEGs) in a particular
  cluster. Volcano plots are used to readily show the DEGs by plotting
  significance versus fold-change on the y and x axes, respectively.
}
\examples{
\donttest{
sc <- DISCBIO(valuesG1msReduced)
sc <- NoiseFiltering(sc, percentile=0.9, CV=0.2, export=FALSE)
sc <- Normalizedata(
    sc, mintotal=1000, minexpr=0, minnumber=0, maxexpr=Inf, downsample=FALSE,
    dsn=1, rseed=17000
)
sc <- FinalPreprocessing(sc, GeneFlitering="NoiseF")
sc <- Clustexp(sc, cln=3) # K-means clustering
sc <- comptSNE(sc, max_iter=100)
dff <- DEGanalysis2clust(sc, Clustering="K-means", K=3, fdr=0.1, name="Name")
name <- dff[[2]][1, 6]
U <- read.csv(file = paste0(name), head=TRUE, sep=",")
VolcanoPlot(U, value=0.05, name=name, adj=FALSE, FS=.4)
}
}
