% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samr-adapted.R
\name{wilcoxon.unpaired.seq.func}
\alias{wilcoxon.unpaired.seq.func}
\title{Twoclass Wilcoxon statistics}
\usage{
wilcoxon.unpaired.seq.func(xresamp, y)
}
\arguments{
\item{xresamp}{an rank array with dim #gene*#sample*nresamp}

\item{y}{outcome vector of values 1 and 2}
}
\value{
the statistic.
}
\description{
Twoclass Wilcoxon statistics
}
