% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-generic-Clustexp.R
\docType{methods}
\name{Clustexp}
\alias{Clustexp}
\alias{Clustexp,PSCANseq-method}
\title{title}
\usage{
Clustexp(
  object,
  clustnr = 20,
  bootnr = 50,
  metric = "pearson",
  do.gap = TRUE,
  SE.method = "Tibs2001SEmax",
  SE.factor = 0.25,
  B.gap = 50,
  cln = 0,
  rseed = 17000
)

\S4method{Clustexp}{PSCANseq}(
  object,
  clustnr = 20,
  bootnr = 50,
  metric = "pearson",
  do.gap = TRUE,
  SE.method = "Tibs2001SEmax",
  SE.factor = 0.25,
  B.gap = 50,
  cln = 0,
  rseed = 17000
)
}
\arguments{
\item{object}{object}

\item{clustnr}{clustnr}

\item{bootnr}{bootnr}

\item{metric}{metric}

\item{do.gap}{do.gap}

\item{SE.method}{SE.method}

\item{SE.factor}{SE.factor}

\item{B.gap}{B.gap}

\item{cln}{cln}

\item{rseed}{rseed}
}
\description{
title
}
