% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-generic-comptSNE.R
\name{comptSNE}
\alias{comptSNE}
\alias{comptSNE,PSCANseq-method}
\title{title}
\usage{
comptSNE(object, rseed = 15555)

\S4method{comptSNE}{PSCANseq}(object, rseed = 15555)
}
\arguments{
\item{object}{object}

\item{rseed}{rseed}
}
\description{
description
}
