% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-generic-comptSNE.R
\name{comptSNE}
\alias{comptSNE}
\alias{comptSNE,DISCBIO-method}
\title{Computing tSNE for K-means clustering}
\usage{
comptSNE(object, rseed = 15555, quiet = FALSE)

\S4method{comptSNE}{DISCBIO}(object, rseed = 15555, quiet = FALSE)
}
\arguments{
\item{object}{\code{DISCBIO} class object.}

\item{rseed}{Integer number. Random seed to to yield exactly reproducible maps across different runs. Default is 15555.}

\item{quiet}{if `TRUE`, suppresses intermediate output}
}
\value{
The DISCBIO-class object input with the tsne slot filled.
}
\description{
This function is used to compute the t-Distributed Stochastic Neighbor Embedding (t-SNE).
}
\examples{
sc <- DISCBIO(valuesG1msReduced) # changes signature of data
sc <- Clustexp(sc, cln=3) # data must be clustered before plottin
sc <- comptSNE(sc, rseed=15555, quiet=TRUE)
head(sc@tsne)
}
