% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-generic-comptsneMB.R
\name{comptsneMB}
\alias{comptsneMB}
\alias{comptsneMB,DISCBIO-method}
\title{Computing tSNE for Model-based clustering}
\usage{
comptsneMB(
  object,
  rseed = NULL,
  max_iter = 5000,
  epoch = 500,
  quiet = FALSE,
  ...
)

\S4method{comptsneMB}{DISCBIO}(
  object,
  rseed = NULL,
  max_iter = 5000,
  epoch = 500,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{object}{\code{DISCBIO} class object.}

\item{rseed}{Random number to be set for reproducible results.}

\item{max_iter}{maximum number of iterations to perform.}

\item{epoch}{The number of iterations in between update messages.}

\item{quiet}{if `TRUE`, suppresses intermediate output.}

\item{...}{other parameters to be passed to `tsne::tsne`.}
}
\value{
The DISCBIO-class object input with the MBtsne slot filled.
}
\description{
This function is used to compute the t-Distributed Stochastic
  Neighbor Embedding (t-SNE).
}
