% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-generic-comptsneMB.R
\name{comptsneMB}
\alias{comptsneMB}
\alias{comptsneMB,PSCANseq-method}
\title{title}
\usage{
comptsneMB(object, rseed = 15555)

\S4method{comptsneMB}{PSCANseq}(object, rseed = 15555)
}
\arguments{
\item{object}{object}

\item{rseed}{rseed}
}
\description{
description
}
