% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGanalysis.R
\name{DEGanalysis}
\alias{DEGanalysis}
\title{title}
\usage{
DEGanalysis(object, Cluster_ID, K, fdr, name)
}
\arguments{
\item{object}{object}

\item{Cluster_ID}{Cluster_ID}

\item{K}{K}

\item{fdr}{fdr}

\item{name}{name}
}
\description{
description
}
