% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-generic-DEGanalysis.R
\name{DEGanalysis}
\alias{DEGanalysis}
\alias{DEGanalysis,DISCBIO-method}
\title{Determining differentially expressed genes (DEGs) between all individual clusters.}
\usage{
DEGanalysis(
  object,
  Clustering = "K-means",
  K,
  fdr = 0.05,
  name = "Name",
  export = TRUE,
  quiet = FALSE,
  plot = TRUE
)

\S4method{DEGanalysis}{DISCBIO}(
  object,
  Clustering = "K-means",
  K,
  fdr = 0.05,
  name = "Name",
  export = TRUE,
  quiet = FALSE,
  plot = TRUE
)
}
\arguments{
\item{object}{\code{DISCBIO} class object.}

\item{Clustering}{Clustering has to be one of the following: ["K-means","MB"]. Default is "K-means"}

\item{K}{A numeric value of the number of clusters.}

\item{fdr}{A numeric value of the false discovery rate. Default is 0.05.}

\item{name}{A string vector showing the name to be used to save the resulted tables.}

\item{export}{A logical vector that allows writing the final gene list in excel file. Default is TRUE.}

\item{quiet}{if `TRUE`, suppresses intermediate text output}

\item{plot}{if `TRUE`, plots are generated}
}
\description{
This function defines DEGs between all individual clusters generated by either K-means or model based clustering.
}
\examples{
\dontrun{
sc <- DISCBIO(valuesG1ms)
sc <- NoiseFiltering(sc, export=FALSE)
sc <- Normalizedata(
    sc, mintotal=1000, minexpr=0, minnumber=0, maxexpr=Inf, downsample=FALSE,
    dsn=1, rseed=17000
)
sc <- Clustexp(sc, cln=3) # K-means clustering
sc <- FinalPreprocessing(sc, GeneFlitering="NoiseF", export=FALSE)
sc <- Clustexp(sc, cln=3) # K-means clustering
sc <- comptSNE(sc, rseed=15555)
DEGanalysis(
    sc, Clustering="K-means", K=3, fdr=0.1, name="Name", export = FALSE
)
}
}
