% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscBIO_class.R
\name{DEGanalysisM}
\alias{DEGanalysisM}
\title{title}
\usage{
DEGanalysisM(object, Cluster_ID, fdr, name, First, Second)
}
\arguments{
\item{object}{object}

\item{Cluster_ID}{Cluster_ID}

\item{fdr}{fdr}

\item{name}{name}

\item{First}{First}

\item{Second}{Second}
}
\description{
description
}
