% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-generic-DEGanalysisM.R
\name{DEGanalysisM}
\alias{DEGanalysisM}
\alias{DEGanalysisM,DISCBIO-method}
\title{Determining differentially expressed genes (DEGs) between two particular clusters.}
\usage{
DEGanalysisM(
  object,
  Clustering = "K-means",
  K,
  fdr = 0.05,
  name = "Name",
  First = "CL1",
  Second = "CL2",
  export = TRUE,
  quiet = FALSE
)

\S4method{DEGanalysisM}{DISCBIO}(
  object,
  Clustering = "K-means",
  K,
  fdr = 0.05,
  name = "Name",
  First = "CL1",
  Second = "CL2",
  export = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{object}{\code{DISCBIO} class object.}

\item{Clustering}{Clustering has to be one of the following: ["K-means","MB"]. Default is "K-means"}

\item{K}{A numeric value of the number of clusters.}

\item{fdr}{A numeric value of the false discovery rate. Default is 0.05.}

\item{name}{A string vector showing the name to be used to save the resulted tables.}

\item{First}{A string vector showing the first target cluster.  Default is "CL1"}

\item{Second}{A string vector showing the second target cluster.  Default is "CL2"}

\item{export}{A logical vector that allows writing the final gene list in excel file. Default is TRUE.}

\item{quiet}{if `TRUE`, suppresses intermediate text output}
}
\description{
This function defines DEGs between particular clusters generated by either K-means or model based clustering.
}
\examples{
sc <- DISCBIO(valuesG1ms)
sc <- NoiseFiltering(sc, plot=FALSE, export=FALSE, quiet=TRUE)
sc <- Normalizedata(
    sc, mintotal=1000, minexpr=0, minnumber=0, maxexpr=Inf, downsample=FALSE
    dsn=1, rseed=17000
)
sc <- Clustexp(sc, cln=3, quiet=TRUE) # K-means clustering
sc <- FinalPreprocessing(sc, GeneFlitering="NoiseF", export=FALSE, quiet=TRUE)
sc <- Clustexp(sc, cln=3, quiet=TRUE) # K-means clustering
sc <- comptSNE(sc, rseed=15555, quiet=TRUE)
DEGanalysisM(
    sc, Clustering="K-means", K=3, fdr=0.1, name="Name", export = FALSE
)
}
