% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exprmclust.R
\name{Exprmclust}
\alias{Exprmclust}
\alias{Exprmclust,PSCANseq-method}
\title{Performing Model-based clustering on expression values}
\usage{
Exprmclust(
  object,
  clusternum = 3,
  modelNames = "VVV",
  reduce = T,
  cluster = NULL,
  quiet = FALSE
)

\S4method{Exprmclust}{PSCANseq}(
  object,
  clusternum = 3,
  modelNames = "VVV",
  reduce = T,
  cluster = NULL,
  quiet = FALSE
)
}
\arguments{
\item{object}{\code{PSCANseq} class object.}

\item{clusternum}{An integer vector specifying all possible cluster numbers. Default is 3.}

\item{modelNames}{model to be used in model-based clustering. By default "ellipsoidal, varying volume, shape, and orientation" is used.}

\item{reduce}{A logical vector that allows performing the PCA on the expression data. Default is TRUE.}

\item{cluster}{A vector showing the ID of cells in the clusters.}

\item{quiet}{if `TRUE`, suppresses intermediary output}
}
\description{
this function first uses principal component analysis (PCA) to reduce dimensionality of original data. It then performs model-based clustering 
on the transformed expression values.
}
