% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscBIO_class.R
\name{FindOutliersKM}
\alias{FindOutliersKM}
\alias{FindOutliersKM,PSCANseq-method}
\title{title}
\usage{
FindOutliersKM(
  object,
  outminc = 5,
  outlg = 2,
  probthr = 0.001,
  thr = 2^-(1:40),
  outdistquant = 0.75
)

\S4method{FindOutliersKM}{PSCANseq}(
  object,
  outminc = 5,
  outlg = 2,
  probthr = 0.001,
  thr = 2^-(1:40),
  outdistquant = 0.75
)
}
\arguments{
\item{object}{object}

\item{outminc}{outminc}

\item{outlg}{outlg}

\item{probthr}{probthr}

\item{thr}{thr}

\item{outdistquant}{outdistquant}
}
\description{
description
}
