% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-generic-FindOutliersKM.R
\name{FindOutliersKM}
\alias{FindOutliersKM}
\alias{FindOutliersKM,PSCANseq-method}
\title{title}
\usage{
FindOutliersKM(
  object,
  K,
  outminc = 5,
  outlg = 2,
  probthr = 0.001,
  thr = 2^-(1:40),
  outdistquant = 0.75,
  plot = TRUE
)

\S4method{FindOutliersKM}{PSCANseq}(
  object,
  K,
  outminc = 5,
  outlg = 2,
  probthr = 0.001,
  thr = 2^-(1:40),
  outdistquant = 0.75,
  plot = TRUE
)
}
\arguments{
\item{object}{object}

\item{K}{K}

\item{outminc}{outminc}

\item{outlg}{outlg}

\item{probthr}{probthr}

\item{thr}{thr}

\item{outdistquant}{outdistquant}

\item{plot}{If `TRUE`, produces a plot of -log10prob per K}
}
\description{
description
}
