% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samr-adapted.R
\name{foldchange.seq.twoclass.unpaired}
\alias{foldchange.seq.twoclass.unpaired}
\title{Foldchange of twoclass unpaired sequencing data}
\usage{
foldchange.seq.twoclass.unpaired(x, y, depth)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{depth}{depth}
}
\description{
Foldchange of twoclass unpaired sequencing data
}
