% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J48DT.R
\name{J48DT}
\alias{J48DT}
\title{J48 Decision Tree}
\usage{
J48DT(data, quiet = FALSE, plot = TRUE)
}
\arguments{
\item{data}{A data frame resulted from running the function ClassVectoringDT.}

\item{quiet}{If `TRUE`, suppresses intermediary output}

\item{plot}{If `FALSE`, suppresses plot output}
}
\value{
Information about the J48 model and, by default, a plot of the
  decision tree.
}
\description{
The decision tree analysis is implemented over a training
  dataset, which consisted of the DEGs obtained by either SAMseq or the
  binomial differential expression.
}
