% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-generic-KMClustDiffGenes.R
\name{KMClustDiffGenes}
\alias{KMClustDiffGenes}
\alias{KMClustDiffGenes,PSCANseq-method}
\title{ClustDiffGenes}
\usage{
KMClustDiffGenes(object, fdr = 0.01)

\S4method{KMClustDiffGenes}{PSCANseq}(object, fdr = 0.01)
}
\arguments{
\item{object}{\code{PSCANseq} class object.}

\item{fdr}{A numeric value of the false discovery rate. Default is 0.01.}
}
\description{
description
}
