% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-generic-KMClustDiffGenes.R
\name{KMClustDiffGenes}
\alias{KMClustDiffGenes}
\alias{KMClustDiffGenes,DISCBIO-method}
\title{ClustDiffGenes}
\usage{
KMClustDiffGenes(object, K, fdr = 0.01, export = TRUE, quiet = FALSE)

\S4method{KMClustDiffGenes}{DISCBIO}(object, K, fdr = 0.01, export = TRUE, quiet = FALSE)
}
\arguments{
\item{object}{\code{DISCBIO} class object.}

\item{K}{A numeric value of the number of clusters.}

\item{fdr}{A numeric value of the false discovery rate. Default is 0.01.}

\item{export}{A logical vector that allows writing the final gene list in excel file. Default is TRUE.}

\item{quiet}{if `TRUE`, suppresses intermediate text output}
}
\description{
description
}
\examples{
\dontrun{
sc <- DISCBIO(valuesG1ms)
sc <- Clustexp(sc, cln=3, quiet=TRUE) # K-means clustering
KMClustDiffGenes(sc, K=3, fdr=.3)
}
}
