% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-generic-KmeanOrder.R
\name{KmeanOrder}
\alias{KmeanOrder}
\alias{KmeanOrder,DISCBIO-method}
\title{Pseudo-time ordering based on k-means clusters}
\usage{
KmeanOrder(object, quiet = FALSE, export = TRUE)

\S4method{KmeanOrder}{DISCBIO}(object, quiet = FALSE, export = TRUE)
}
\arguments{
\item{object}{\code{DISCBIO} class object.}

\item{quiet}{if `TRUE`, suppresses intermediary output}

\item{export}{if `TRUE`, exports order table to csv}
}
\value{
The DISCBIO-class object input with the kordering slot filled.
}
\description{
This function takes the exact output of exprmclust function and
  construct Pseudo-time ordering by mapping all cells onto the path that
  connects cluster centers.
}
\examples{
data(valuesG1msReduced_treated_K)  # DIscBIO:::prepExampleDataset for details
sc <- valuesG1msReduced_treated_K
Order <- KmeanOrder(sc, export = FALSE)
Order@kordering
}
