% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscBIO_class.R
\name{KmeanOrder}
\alias{KmeanOrder}
\title{title}
\usage{
KmeanOrder(object, Clusters, sampleNames)
}
\arguments{
\item{object}{object}

\item{Clusters}{Clusters}

\item{sampleNames}{sampleNames}
}
\description{
description
}
