% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-generic-KmeanOrder.R
\name{KmeanOrder}
\alias{KmeanOrder}
\alias{KmeanOrder,PSCANseq-method}
\title{Pseudo-time ordering based on k-means clusters}
\usage{
KmeanOrder(object, quiet = FALSE, export = TRUE)

\S4method{KmeanOrder}{PSCANseq}(object, quiet = FALSE, export = TRUE)
}
\arguments{
\item{object}{\code{PSCANseq} class object.}

\item{quiet}{if `TRUE`, suppresses intermediary output}

\item{export}{if `TRUE`, exports order table to csv}
}
\description{
This function takes the exact output of exprmclust function and construct Pseudo-time ordering by mapping all cells onto the path that connects cluster centers.
}
