% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kmeansruns.R
\name{Kmeansruns}
\alias{Kmeansruns}
\title{K-means runs}
\usage{
Kmeansruns(
  data,
  krange = 2:10,
  criterion = "ch",
  iter.max = 100,
  runs = 100,
  scaledata = FALSE,
  alpha = 0.001,
  critout = FALSE,
  plot = FALSE,
  method = "euclidean",
  ...
)
}
\arguments{
\item{data}{data}

\item{krange}{krange}

\item{criterion}{criterion}

\item{iter.max}{iter.max}

\item{runs}{number of runs}

\item{scaledata}{scaledata}

\item{alpha}{alpha}

\item{critout}{critout}

\item{plot}{plot}

\item{method}{method}

\item{...}{extra parameters to be passed to kmm}
}
\description{
K-means runs
}
