% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MB_Order.R
\name{MB_Order}
\alias{MB_Order}
\title{title}
\usage{
MB_Order(object, sampleNames, Names, quiet = FALSE, export = TRUE)
}
\arguments{
\item{object}{object}

\item{sampleNames}{sampleNames}

\item{Names}{Names}

\item{quiet}{if `TRUE`, intermediary output is suppressed}

\item{export}{if `TRUE`, exports the results as a CSV file}
}
\description{
description
}
