% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MB_Order.R
\name{MB_Order}
\alias{MB_Order}
\title{Pseudo-time ordering based on Model-based clusters}
\usage{
MB_Order(
  object,
  quiet = FALSE,
  export = FALSE,
  filename = "Cellular_pseudo-time_ordering_based_on_Model-based_clusters"
)
}
\arguments{
\item{object}{\code{DISCBIO} class object.}

\item{quiet}{if `TRUE`, intermediary output is suppressed}

\item{export}{if `TRUE`, exports the results as a CSV file}

\item{filename}{Name of the exported file (if `export=TRUE`)}
}
\value{
The DISCBIO-class object input with the MBordering slot filled.
}
\description{
This function takes the exact output of exprmclust function and
  construct Pseudo-time ordering by mapping all cells onto the path that
  connects cluster centers.
}
