% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-generic-MBClustDiffGenes.R
\name{MBClustDiffGenes}
\alias{MBClustDiffGenes}
\alias{MBClustDiffGenes,PSCANseq-method}
\title{ClustDiffGenes}
\usage{
MBClustDiffGenes(object, fdr = 0.01)

\S4method{MBClustDiffGenes}{PSCANseq}(object, fdr = 0.01)
}
\arguments{
\item{object}{\code{PSCANseq} class object.}

\item{fdr}{A numeric value of the false discovery rate. Default is 0.01.}
}
\description{
ClustDiffGenes
}
