% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetAnalysis.R
\name{NetAnalysis}
\alias{NetAnalysis}
\title{Networking analysis.}
\usage{
NetAnalysis(data, export = TRUE, FileName = "1")
}
\arguments{
\item{data}{Protein-protein interaction data frame resulted from running the
PPI function.}

\item{export}{if `TRUE`, exports the analysis table as a csv file}

\item{FileName}{suffix for the file name (if export = TRUE)}
}
\value{
A network analysis table
}
\description{
This function checks the connectivity degree and the betweenness
  centrality, which reflect the communication flow in the defined PPI
  networks
}
\examples{
\donttest{
sc <- DISCBIO(valuesG1msReduced)
sc <- NoiseFiltering(sc, percentile=0.9, CV=0.2, export=FALSE)
sc <- Normalizedata(
    sc, mintotal=1000, minexpr=0, minnumber=0, maxexpr=Inf, downsample=FALSE,
    dsn=1, rseed=17000
)
sc <- FinalPreprocessing(sc, GeneFlitering="NoiseF")
sc <- Clustexp(sc, cln=3) # K-means clustering
sc <- comptSNE(sc, max_iter=100)
dff <- DEGanalysis2clust(sc, Clustering="K-means", K=3, fdr=0.1, name="Name")
DEGs <- dff[[2]][1, 6]
data <- read.csv(file=paste0(DEGs),head=TRUE,sep=",")
data <- data[,3]
FileName <- paste0(DEGs)
ppi <- PPI(data, FileName)
NetAnalysis(ppi)
}
}
