% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetAnalysis.R
\name{NetAnalysis}
\alias{NetAnalysis}
\title{Networking analysis.}
\usage{
NetAnalysis(data, export = FALSE, FileName = "NetworkAnalysisTable-1")
}
\arguments{
\item{data}{Protein-protein interaction data frame resulted from running the
PPI function.}

\item{export}{if `TRUE`, exports the analysis table as a csv file}

\item{FileName}{suffix for the file name (if export = TRUE)}
}
\value{
A network analysis table
}
\description{
This function checks the connectivity degree and the betweenness
  centrality, which reflect the communication flow in the defined PPI
  networks
}
