% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networking.R
\name{Networking}
\alias{Networking}
\title{Plotting the network.}
\usage{
Networking(
  data,
  FileName = NULL,
  species = "9606",
  plot_width = 25,
  plot_height = 15
)
}
\arguments{
\item{data}{A gene list.}

\item{FileName}{A string vector showing the name to be used to save the
resulted network. If `NULL`, the network will be saved to a temporary
directory}

\item{species}{The taxonomy name/id. Default is "9606" for Homo sapiens.}

\item{plot_width}{Plot width}

\item{plot_height}{Plot height}
}
\value{
A plot of the network
}
\description{
This function uses STRING-api to plot the network.
}
\examples{
\donttest{
sc <- DISCBIO(valuesG1msReduced)
sc <- NoiseFiltering(sc, percentile=0.9, CV=0.2, export=FALSE, plot=FALSE)
sc <- Normalizedata(
    sc, mintotal=1000, minexpr=0, minnumber=0, maxexpr=Inf, downsample=FALSE,
    dsn=1, rseed=17000
)
sc <- FinalPreprocessing(sc, export=FALSE, quiet=TRUE)
sc <- Clustexp(sc, cln=3, quiet=TRUE) #' K-means clustering
sc <- comptSNE(sc, max_iter=100, quiet=TRUE)
dff <- DEGanalysis2clust(sc, K=3, export=FALSE, quiet=TRUE, plot=FALSE)
Networking(dff$FinalDEGsU[, "Gene Name"])
}
}
