% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NoiseFiltering.R
\name{NoiseFiltering}
\alias{NoiseFiltering}
\title{Noise Filtering}
\usage{
NoiseFiltering(
  object,
  percentile,
  CV,
  GeneList,
  geneCol,
  FgeneCol,
  erccCol,
  Val = TRUE,
  plot = TRUE,
  export = TRUE
)
}
\arguments{
\item{object}{object}

\item{percentile}{percentile}

\item{CV}{CV}

\item{GeneList}{GeneList}

\item{geneCol}{geneCol}

\item{FgeneCol}{FgeneCol}

\item{erccCol}{erccCol}

\item{Val}{Val}

\item{plot}{Generate plot?}

\item{export}{Export results?}
}
\description{
Given a matrix or data frame of count data, this function 
estimates the size factors as follows: Each column is divided by the 
geometric means of the rows. The median (or, if requested, another location
estimator) of these ratios (skipping the genes with a # geometric mean of
zero) is used as the size factor for this column. Source: DESeq package.
}
