% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-generic-PCAplotSymbols.R
\name{PCAplotSymbols}
\alias{PCAplotSymbols}
\alias{PCAplotSymbols,DISCBIO-method}
\title{Plot PCA symbols}
\usage{
PCAplotSymbols(object, types = NULL)

\S4method{PCAplotSymbols}{DISCBIO}(object, types = NULL)
}
\arguments{
\item{object}{\code{DISCBIO} class object.}

\item{types}{If types=NULL then the names of the cells will be grouped
automatically. Default is NULL}
}
\value{
Plot of the Principal Components
}
\description{
Generates a plot of grouped PCA components
}
\examples{
sc <- DISCBIO(valuesG1msReduced)
sc <- NoiseFiltering(sc, percentile=0.9, CV=0.2, export=FALSE)
sc <- Normalizedata(
    sc, mintotal=1000, minexpr=0, minnumber=0, maxexpr=Inf, downsample=FALSE,
    dsn=1, rseed=17000
)
sc <- FinalPreprocessing(sc, GeneFlitering="NoiseF", export=FALSE)
sc <- Exprmclust(sc, K=2)
PCAplotSymbols(sc)
}
