% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAplotSymbols.R
\name{PCAplotSymbols}
\alias{PCAplotSymbols}
\title{title}
\usage{
PCAplotSymbols(object1, object2, types = NULL)
}
\arguments{
\item{object1}{object1}

\item{object2}{object2}

\item{types}{types}
}
\description{
description
}
