% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-generic-PCAplotSymbols.R
\name{PCAplotSymbols}
\alias{PCAplotSymbols}
\alias{PCAplotSymbols,DISCBIO-method}
\title{Plot PCA symbols}
\usage{
PCAplotSymbols(object, types = NULL)

\S4method{PCAplotSymbols}{DISCBIO}(object, types = NULL)
}
\arguments{
\item{object}{\code{DISCBIO} class object.}

\item{types}{If types=NULL then the names of the cells will be grouped
automatically. Default is NULL}
}
\value{
Plot of the Principal Components
}
\description{
Generates a plot of grouped PCA components
}
