% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-generic-plotexptsneMB.R
\name{plotexptsneMB}
\alias{plotexptsneMB}
\alias{plotexptsneMB,DISCBIO-method}
\title{Highlighting gene expression in Model-based clustering in the t-SNE map}
\usage{
plotexptsneMB(object, g, n = NULL)

\S4method{plotexptsneMB}{DISCBIO}(object, g, n = NULL)
}
\arguments{
\item{object}{\code{DISCBIO} class object.}

\item{g}{Individual gene name or vector with a group of gene names corresponding to a subset of valid row names of the \code{ndata} slot
of the \code{DISCBIO} object.}

\item{n}{String of characters representing the title of the plot. Default is NULL and the first element of \code{g} is chosen.}
}
\description{
The t-SNE map representation can also be used to analyze expression of a gene or a group of genes, 
to investigate cluster specific gene expression patterns
}
\examples{
\dontrun{
sc <- DISCBIO(valuesG1msReduced)
sc <- NoiseFiltering(sc, export=FALSE)
sc <- Normalizedata(
    sc, mintotal=1000, minexpr=0, minnumber=0, maxexpr=Inf, downsample=FALSE,
    dsn=1, rseed=17000
)
sc <- FinalPreprocessing(sc, GeneFlitering="NoiseF", export=FALSE)
sc <- Exprmclust(sc)
sc <- comptsneMB(sc, rseed=15555)
sc <- Clustexp(sc, cln=3)
sc <- MB_Order(sc, export = FALSE)
g <- 'ENSG00000001460'
plotexptsneMB(sc, g)
}
}
