% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-generic-plotGap.R
\name{plotGap}
\alias{plotGap}
\alias{plotGap,DISCBIO-method}
\title{Plotting Gap Statistics}
\usage{
plotGap(object)

\S4method{plotGap}{DISCBIO}(object)
}
\arguments{
\item{object}{\code{DISCBIO} class object.}
}
\description{
Plotting Gap Statistics
}
\examples{
sc <- DISCBIO(valuesG1ms) # changes signature of data
sc <- Clustexp(sc, cln=3) # data must be clustered before plotting
plotGap(sc)
}
