% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-generic-plotGap.R
\name{plotGap}
\alias{plotGap}
\alias{plotGap,PSCANseq-method}
\title{Plotting Gap Statistics}
\usage{
plotGap(object)

\S4method{plotGap}{PSCANseq}(object)
}
\arguments{
\item{object}{\code{PSCANseq} class object.}
}
\description{
Plotting Gap Statistics
}
