% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-generic-plotGap.R
\name{plotGap}
\alias{plotGap}
\alias{plotGap,DISCBIO-method}
\title{Plotting Gap Statistics}
\usage{
plotGap(object, y_limits = NULL)

\S4method{plotGap}{DISCBIO}(object, y_limits = NULL)
}
\arguments{
\item{object}{\code{DISCBIO} class object.}

\item{y_limits}{2-length numeric vector with the limits for the gap plot}
}
\value{
A plot of the gap statistics
}
\description{
Plotting Gap Statistics
}
