% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-generic-plotKmeansLabelstSNE.R
\name{plotKmeansLabelstSNE}
\alias{plotKmeansLabelstSNE}
\alias{plotKmeansLabelstSNE,DISCBIO-method}
\title{tSNE map for K-means clustering with labels}
\usage{
plotKmeansLabelstSNE(object)

\S4method{plotKmeansLabelstSNE}{DISCBIO}(object)
}
\arguments{
\item{object}{\code{DISCBIO} class object.}
}
\value{
Plot containing the ID of the cells in each cluster
}
\description{
Visualizing the K-means clusters using tSNE maps
}
\examples{
data(valuesG1msReduced_treated_K)  # DIscBIO:::prepExampleDataset for details
sc <- valuesG1msReduced_treated_K
plotKmeansLabelstSNE(sc) # Plots the ID of the cells in each cluster
}
