% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-generic-plotKmeansLabelstSNE.R
\name{plotKmeansLabelstSNE}
\alias{plotKmeansLabelstSNE}
\title{tSNE map for K-means clustering with labels}
\usage{
plotKmeansLabelstSNE(object)
}
\arguments{
\item{object}{\code{DISCBIO} class object.}
}
\description{
Visualizing the K-means clusters using tSNE maps
}
