% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-generic-plotMBLabelstSNE.R
\name{plotMBLabelstSNE}
\alias{plotMBLabelstSNE}
\alias{plotMBLabelstSNE,DISCBIO-method}
\title{title tSNE map for Model-based clustering with labels}
\usage{
plotMBLabelstSNE(object)

\S4method{plotMBLabelstSNE}{DISCBIO}(object)
}
\arguments{
\item{object}{\code{DISCBIO} class object.}
}
\description{
Visualizing the Model-based clusters using tSNE maps
}
\examples{
\dontrun{
sc <- DISCBIO(valuesG1msReduced)
sc <- NoiseFiltering(sc, export=FALSE)
sc <- Normalizedata(
    sc, mintotal=1000, minexpr=0, minnumber=0, maxexpr=Inf, downsample=FALSE,
    dsn=1, rseed=17000
)
sc <- FinalPreprocessing(sc, GeneFlitering="NoiseF", export=FALSE)
sc <- Exprmclust(sc)
sc <- comptsneMB(sc, rseed=15555, quiet = TRUE)
plotMBLabelstSNE(sc)
}
}
