% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-generic-plotMBLabelstSNE.R
\name{plotMBLabelstSNE}
\alias{plotMBLabelstSNE}
\alias{plotMBLabelstSNE,PSCANseq-method}
\title{title tSNE map for Model-based clustering with labels}
\usage{
plotMBLabelstSNE(object)

\S4method{plotMBLabelstSNE}{PSCANseq}(object)
}
\arguments{
\item{object}{\code{PSCANseq} class object.}
}
\description{
Visualizing the Model-based clusters using tSNE maps
}
