% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMBorderPCA.R
\name{PlotMBorderPCA}
\alias{PlotMBorderPCA}
\title{Plotting pseudo-time ordering in Model-based clustering in PCA.}
\usage{
PlotMBorderPCA(object)
}
\arguments{
\item{object}{\code{PSCANseq} class object.}
}
\description{
The PCA representation can also be used to show the pseudo-time ordering.
}
