% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMBorderPCA.R
\name{PlotMBorderPCA}
\alias{PlotMBorderPCA}
\title{Plotting pseudo-time ordering in Model-based clustering in PCA.}
\usage{
PlotMBorderPCA(object)
}
\arguments{
\item{object}{\code{DISCBIO} class object.}
}
\description{
The PCA representation can also be used to show the pseudo-time ordering.
}
\examples{
sc <- DISCBIO(valuesG1msReduced)
sc <- NoiseFiltering(sc, percentile=0.9, CV=0.2, export=FALSE)
sc <- Normalizedata(
    sc, mintotal=1000, minexpr=0, minnumber=0, maxexpr=Inf, downsample=FALSE,
    dsn=1, rseed=17000
)
sc <- FinalPreprocessing(sc, GeneFlitering="NoiseF", export=FALSE)
sc <- Exprmclust(sc, K=3)
sc <- comptsneMB(sc, rseed=15555)
sc <- Clustexp(sc, cln=3)
sc <- MB_Order(sc, export = FALSE)
PlotMBorderPCA(sc)
}
