% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-generic-PlotmclustMB.R
\name{PlotmclustMB}
\alias{PlotmclustMB}
\alias{PlotmclustMB,PSCANseq-method}
\title{Plotting the Model-based clusters in PCA.}
\usage{
PlotmclustMB(object)

\S4method{PlotmclustMB}{PSCANseq}(object)
}
\arguments{
\item{object}{\code{PSCANseq} class object.}
}
\description{
Plot the model-based clustering results
}
