% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-generic-PlotmclustMB.R
\name{PlotmclustMB}
\alias{PlotmclustMB}
\alias{PlotmclustMB,DISCBIO-method}
\title{Plotting the Model-based clusters in PCA.}
\usage{
PlotmclustMB(object)

\S4method{PlotmclustMB}{DISCBIO}(object)
}
\arguments{
\item{object}{\code{DISCBIO} class object.}
}
\value{
A plot of the PCA.
}
\description{
Plot the model-based clustering results
}
