% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-generic-plotOrderMBtsne.R
\name{plotOrderMBtsne}
\alias{plotOrderMBtsne}
\alias{plotOrderMBtsne,PSCANseq-method}
\title{Plotting the pseudo-time ordering based on Model-based clusters in the t-SNE map}
\usage{
plotOrderMBtsne(object)

\S4method{plotOrderMBtsne}{PSCANseq}(object)
}
\arguments{
\item{object}{\code{PSCANseq} class object.
of the \code{PSCANseq} object.}
}
\description{
The tSNE representation can also be used to show the pseudo-time ordering.
}
