% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscBIO_class.R
\name{plotSilhouette}
\alias{plotSilhouette}
\alias{plotSilhouette,PSCANseq-method}
\alias{plottSNE,PSCANseq-method}
\title{title}
\usage{
plotSilhouette(object, K)

\S4method{plotSilhouette}{PSCANseq}(object)

\S4method{plottSNE}{PSCANseq}(object)
}
\arguments{
\item{object}{object}

\item{K}{K}
}
\description{
description

description
}
