% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-generic-plotSilhouette.R,
%   R/DIscBIO-generic-plottSNE.R
\name{plotSilhouette}
\alias{plotSilhouette}
\alias{plotSilhouette,PSCANseq-method}
\alias{plottSNE,PSCANseq-method}
\title{Silhouette Plot for K-means clustering}
\usage{
plotSilhouette(object, K)

\S4method{plotSilhouette}{PSCANseq}(object, K)

\S4method{plottSNE}{PSCANseq}(object)
}
\arguments{
\item{object}{\code{PSCANseq} class object.}

\item{K}{A numeric value of the number of clusters}
}
\description{
The silhouette provides a representation of how well each point is represented by its cluster in comparison to the 
closest neighboring cluster. It computes for each point the difference between the average similarity 
to all points in the same cluster and to all points in the closest neighboring cluster. This difference it 
normalize such that it can take values between -1 and 1 with higher values reflecting better 
representation of a point by its cluster.
}
