% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-generic-plotSilhouette.R,
%   R/DIscBIO-generic-plottSNE.R
\name{plotSilhouette}
\alias{plotSilhouette}
\alias{plotSilhouette,DISCBIO-method}
\alias{plottSNE,DISCBIO-method}
\title{Silhouette Plot for K-means clustering}
\usage{
plotSilhouette(object, K)

\S4method{plotSilhouette}{DISCBIO}(object, K)

\S4method{plottSNE}{DISCBIO}(object)
}
\arguments{
\item{object}{\code{DISCBIO} class object.}

\item{K}{A numeric value of the number of clusters}
}
\value{
A silhouette plot
}
\description{
The silhouette provides a representation of how well each point
  is represented by its cluster in comparison to the closest neighboring
  cluster. It computes for each point the difference between the average
  similarity to all points in the same cluster and to all points in the
  closest neighboring cluster. This difference it normalize such that it can
  take values between -1 and 1 with higher values reflecting better
  representation of a point by its cluster.
}
\examples{
sc <- DISCBIO(valuesG1msReduced) # changes signature of data
sc <- Clustexp(sc, cln=3) # data must be clustered before plottin
sc <- comptSNE(sc, max_iter=100)
plotSilhouette(sc, K=3)
}
