% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-classes.R,
%   R/DIscBIO-generic-plotsilhouetteMB.R
\name{plotsilhouetteMB}
\alias{plotsilhouetteMB}
\alias{plotsilhouetteMB,PSCANseq-method}
\title{title}
\usage{
plotsilhouetteMB(object, K)

\S4method{plotsilhouetteMB}{PSCANseq}(object)

plotsilhouetteMB(object, K)

\S4method{plotsilhouetteMB}{PSCANseq}(object)
}
\arguments{
\item{object}{object}

\item{K}{k}
}
\description{
description

description
}
