% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-generic-plotSymbolstSNE.R
\name{plotSymbolstSNE}
\alias{plotSymbolstSNE}
\alias{plotSymbolstSNE,PSCANseq-method}
\title{tSNE map for K-means clustering with symbols}
\usage{
plotSymbolstSNE(object, types = NULL)

\S4method{plotSymbolstSNE}{PSCANseq}(object, types = NULL)
}
\arguments{
\item{object}{\code{PSCANseq} class object.}

\item{types}{If types=NULL then the names of the cells will be grouped automatically. Default is NULL}
}
\description{
Visualizing the K-means clusters using tSNE maps
}
