% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-generic-plotSymbolstSNE.R
\name{plotSymbolstSNE}
\alias{plotSymbolstSNE}
\alias{plotSymbolstSNE,PSCANseq-method}
\title{title}
\usage{
plotSymbolstSNE(object, types = NULL)

\S4method{plotSymbolstSNE}{PSCANseq}(object, types = NULL)
}
\arguments{
\item{object}{object}

\item{types}{types}
}
\description{
title
}
