% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-generic-plottSNE.R
\name{plottSNE}
\alias{plottSNE}
\title{tSNE map for K-means clustering}
\usage{
plottSNE(object)
}
\arguments{
\item{object}{\code{DISCBIO} class object.}
}
\value{
A plot of t-SNEs.
}
\description{
Visualizing the K-means clusters using tSNE maps
}
\examples{
sc <- DISCBIO(valuesG1msReduced) # changes signature of data
sc <- Clustexp(sc, cln=3) # data must be clustered before plotting
sc <- comptSNE(sc, rseed=15555, quiet=TRUE)
plottSNE(sc)
}
