% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-generic-plottSNE.R
\name{plottSNE}
\alias{plottSNE}
\title{tSNE map for K-means clustering}
\usage{
plottSNE(object)
}
\arguments{
\item{object}{\code{DISCBIO} class object.}
}
\value{
A plot of t-SNEs.
}
\description{
Visualizing the K-means clusters using tSNE maps
}
\examples{
data(valuesG1msReduced_treated_K)  # details: DIscBIO:::prepExampleDataset
sc <- valuesG1msReduced_treated_K
plottSNE(sc)
}
