% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-generic-plottSNE.R
\name{plottSNE}
\alias{plottSNE}
\alias{plottSNE,DISCBIO-method}
\title{tSNE map}
\usage{
plottSNE(object)

\S4method{plottSNE}{DISCBIO}(object)
}
\arguments{
\item{object}{\code{DISCBIO} class object.}
}
\value{
A plot of t-SNEs.
}
\description{
Visualizing the k-means or model-based clusters using tSNE maps
}
