% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-classes.R,
%   R/DIscBIO-generic-plottsneMB.R
\name{plottsneMB}
\alias{plottsneMB}
\alias{plottsneMB,PSCANseq-method}
\title{title}
\usage{
plottsneMB(object, K)

\S4method{plottsneMB}{PSCANseq}(object)

plottsneMB(object, K)

\S4method{plottsneMB}{PSCANseq}(object)
}
\arguments{
\item{object}{object}

\item{K}{K}
}
\description{
description

description
}
