% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-generic-plottsneMB.R
\name{plottsneMB}
\alias{plottsneMB}
\alias{plottsneMB,DISCBIO-method}
\title{tSNE map for Model-based clustering}
\usage{
plottsneMB(object, K = length(table(object@MBclusters$clusterid)))

\S4method{plottsneMB}{DISCBIO}(object, K = length(table(object@MBclusters$clusterid)))
}
\arguments{
\item{object}{\code{DISCBIO} class object.}

\item{K}{A numeric value of the number of clusters}
}
\value{
A plot of t-SNEs.
}
\description{
Visualizing the Model-based clusters using tSNE maps
}
\examples{
\dontrun{
sc <- DISCBIO(valuesG1msReduced)
sc <- NoiseFiltering(sc, percentile=0.9, CV=0.2, export=FALSE)
sc <- Normalizedata(
    sc, mintotal=1000, minexpr=0, minnumber=0, maxexpr=Inf, downsample=FALSE,
    dsn=1, rseed=17000
)
sc <- FinalPreprocessing(sc, GeneFlitering="NoiseF", export=FALSE)
sc <- Exprmclust(sc, K=2)
sc <- comptsneMB(sc, rseed=15555, quiet = TRUE)
plottsneMB(sc)
}
}
