% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPI.R
\name{PPI}
\alias{PPI}
\title{Defining protein-protein interactions (PPI) over a list of genes,}
\usage{
PPI(data, FileName = NULL, species = "9606")
}
\arguments{
\item{data}{A gene list.}

\item{FileName}{A string vector showing the name to be used to save the
resulted table. If null, no file will be exported}

\item{species}{The taxonomy name/id. Default is "9606" for Homo sapiens.}
}
\value{
Either CSV files stored in the user's file system and its
corresponding `data.frame` object in R or and R object containing that
information.
}
\description{
This function uses STRING-api. The outcome of STRING analysis
  will be stored in comma-separated values files.
}
