% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepExampleDataset.R
\name{prepExampleDataset}
\alias{prepExampleDataset}
\title{Prepare Example Dataset}
\usage{
prepExampleDataset(dataset, save = TRUE)
}
\arguments{
\item{dataset}{Dataset used for transformation}

\item{save}{save results?}
}
\value{
rda file
}
\description{
Internal function that prepares a pre-treated dataset for use in
several examples
}
\examples{
\donttest{
    DIscBIO:::prepExampleDataset(valuesG1msReduced, save=FALSE)
}
}
