% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepExampleDataset.R
\name{prepExampleDataset}
\alias{prepExampleDataset}
\title{Prepare Example Dataset}
\usage{
prepExampleDataset(dataset, save = TRUE)
}
\arguments{
\item{dataset}{Dataset used for transformation}

\item{save}{save results?}
}
\value{
Two rda files, ones for K-means clustering and another for
Model-based clustering.
}
\description{
Internal function that prepares a pre-treated dataset for use in
several examples
}
\details{
This function serves the purpose of treating datasets such as
valuesG1msReduced to reduce examples of other functions by bypassing some
analysis steps covered in the vignettes.
}
\author{
Waldir Leoncio
}
