% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscBIO_class.R
\docType{class}
\name{PSCANseq}
\alias{PSCANseq}
\alias{PSCANseq-class,}
\alias{PSCANseq-class}
\title{PSCANseq}
\description{
PSCANseq
}
\section{Slots}{

\describe{
\item{\code{expdata}}{}

\item{\code{ndata}}{}

\item{\code{fdata}}{}

\item{\code{distances}}{}

\item{\code{tsne}}{}

\item{\code{background}}{}

\item{\code{out}}{}

\item{\code{cpart}}{}

\item{\code{fcol}}{}

\item{\code{filterpar}}{}

\item{\code{clusterpar}}{}

\item{\code{outlierpar}}{}

\item{\code{kmeans}}{}

\item{\code{MBclusters}}{}

\item{\code{kordering}}{}

\item{\code{MBordering}}{}

\item{\code{MBtsne}}{}
}}

