% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIscBIO-classes.R
\docType{class}
\name{PSCANseq}
\alias{PSCANseq}
\alias{PSCANseq-class,}
\alias{PSCANseq-class}
\title{The PSCANseq Class}
\arguments{
\item{object}{An PSCANseq object.}
}
\description{
The PSCANseq class is the central object storing all information generated throughout the pipeline.
}
\details{
PSCANseq
}
