% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-functions-samr-adapted.R
\name{rankcols}
\alias{rankcols}
\title{Rank columns}
\usage{
rankcols(x)
}
\arguments{
\item{x}{x}
}
\description{
Ranks the elements within each col of the matrix x and returns
these ranks in a matrix
}
\note{
this function is equivalent to `samr::rankcol`, but uses `apply` to
rank the colums instead of a compiled Fortran function which was causing our
DEGanalysis functions to freeze in large datasets.
}
