% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceDecimals.R
\name{replaceDecimals}
\alias{replaceDecimals}
\title{Replace Decimals}
\usage{
replaceDecimals(x, from = ",", to = ".")
}
\arguments{
\item{x}{vector of characters}

\item{from}{decimal separator on input file}

\item{to}{decimal separator for output file}
}
\description{
Replaces decimals separators between comma and periods on a
character vector
}
\note{
This function was especially designed to be used with retormatSiggenes
}
\seealso{
reformatSiggenes
}
