% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RpartDT.R
\name{RpartDT}
\alias{RpartDT}
\title{RPART Decision Tree}
\usage{
RpartDT(data, quiet = FALSE, plot = TRUE)
}
\arguments{
\item{data}{The exact output of the exprmclust function.}

\item{quiet}{If `TRUE`, suppresses intermediary output}

\item{plot}{If `FALSE`, suppresses plot output}
}
\value{
Information about the model and, by default, a plot of the decision
  tree.
}
\description{
The decision tree analysis is implemented over a training
  dataset, which consisted of the DEGs obtained by either SAMseq or the
  binomial differential expression.
}
\examples{
data(valuesG1msReduced_treated_K)  # details: DIscBIO:::prepExampleDataset
sc <- valuesG1msReduced_treated_K
cdiff <- DEGanalysis2clust(
    sc, Clustering="K-means", K=3, fdr=.2, name="Name", First="CL1",
    Second="CL2", export=FALSE
)
sigDEG <- cdiff[[1]]
DATAforDT <- ClassVectoringDT(
    sc, Clustering="K-means", K=3, First="CL1", Second="CL2", sigDEG,
)
RpartDT(DATAforDT)
}
