% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RpartDT.R
\name{RpartDT}
\alias{RpartDT}
\title{RPART Decision Tree}
\usage{
RpartDT(data)
}
\arguments{
\item{data}{The exact output of the exprmclust function.}
}
\description{
The decision tree analysis is implemented over a training dataset, which consisted of the DEGs obtained by either SAMseq or the binomial differential expression.
}
