% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RpartDT.R
\name{RpartDT}
\alias{RpartDT}
\title{RPART Decision Tree}
\usage{
RpartDT(data, quiet = FALSE, plot = TRUE)
}
\arguments{
\item{data}{The exact output of the exprmclust function.}

\item{quiet}{If `TRUE`, suppresses intermediary output}

\item{plot}{If `FALSE`, suppresses plot output}
}
\description{
The decision tree analysis is implemented over a training dataset, which consisted of the DEGs obtained by either SAMseq or the binomial differential expression.
}
\examples{
sc <- DISCBIO(valuesG1ms)
sc <- NoiseFiltering(sc, export=FALSE)
sc <- Normalizedata(
    sc, mintotal=1000, minexpr=0, minnumber=0, maxexpr=Inf, downsample=FALSE,
    dsn=1, rseed=17000
)
sc <- FinalPreprocessing(sc, GeneFlitering="NoiseF", export=FALSE)
sc <- Clustexp(sc, cln=3) # K-means clustering
sc <- comptSNE(sc, rseed=15555)
cdiff <- DEGanalysis2clust(
    sc, Clustering="K-means", K=3, fdr=.2, name="Name", First="CL1",
    Second="CL2", export=FALSE
)
sigDEG <- cdiff[[1]]
DATAforDT <- ClassVectoringDT(
    sc, Clustering="K-means", K=3, First="CL1", Second="CL2", sigDEG,
)
RpartDT(DATAforDT)
}
